+++ tests/testthat/test-rownames.R
@@ -9,8 +9,14 @@ test_that("has_rownames and remove_rownames", {
 })
 
 test_that("setting row names on a tibble raises a warning", {
-  mtcars_tbl <- as_tibble(mtcars)
-  expect_warning(rownames(mtcars_tbl) <- rownames(mtcars), "deprecated")
+  mtcars2 <- as_tibble(mtcars)
+  expect_false(has_rownames(mtcars2))
+
+  expect_warning(
+    rownames(mtcars2) <- rownames(mtcars),
+    "deprecated",
+    fixed = TRUE
+  )
 })
 
 test_that("rownames_to_column keeps the tbl classes (#882)", {
@@ -24,12 +30,14 @@ test_that("rownames_to_column keeps the tbl classes (#882)", {
     fixed = TRUE
   )
 
-  res1 <- rownames_to_column(as_tibble(mtcars), "Make&Model")
+  mtcars2 <- as_tibble(mtcars, rownames = NA)
+
+  res1 <- rownames_to_column(mtcars2, "Make&Model")
   expect_false(has_rownames(res1))
-  expect_equal(class(res1), class(as_tibble(mtcars)))
+  expect_equal(class(res1), class(mtcars2))
   expect_equal(res1$`Make&Model`, rownames(mtcars))
   expect_error(
-    rownames_to_column(as_tibble(mtcars), "wt"),
+    rownames_to_column(mtcars2, "wt"),
     error_existing_names("wt"),
     fixed = TRUE
   )
@@ -46,12 +54,14 @@ test_that("rowid_to_column keeps the tbl classes", {
     fixed = TRUE
   )
 
-  res1 <- rowid_to_column(as_tibble(mtcars), "row_id")
+  mtcars2 <- as_tibble(mtcars, rownames = NA)
+
+  res1 <- rowid_to_column(mtcars2, "row_id")
   expect_false(has_rownames(res1))
-  expect_equal(class(res1), class(as_tibble(mtcars)))
+  expect_equal(class(res1), class(mtcars2))
   expect_equal(res1$row_id, seq_len(nrow(mtcars)))
   expect_error(
-    rowid_to_column(as_tibble(mtcars), "wt"),
+    rowid_to_column(mtcars2, "wt"),
     error_existing_names("wt"),
     fixed = TRUE
   )
@@ -59,7 +69,8 @@ test_that("rowid_to_column keeps the tbl classes", {
 
 test_that("column_to_rownames returns tbl", {
   var <- "car"
-  mtcars1 <- as_tibble(mtcars)
+  mtcars1 <- as_tibble(mtcars, rownames = NA)
+
   expect_true(has_rownames(mtcars1))
   res0 <- rownames_to_column(mtcars1, var)
   expect_warning(res <- column_to_rownames(res0, var))